Taglet v0.3.0 Taglet

Documentation for Taglet.

Types

context()

context() :: String.t

persisted_struct()

persisted_struct() :: struct

tag()

tag() :: String.t

tag_list()

tag_list() :: list

tags()

tags() :: String.t | list

Functions

add(struct, tags, context \\ "tags")

add(struct, tags, context) :: struct

Get a persisted struct and inserts a new tag associated to this struct for a specific context.

You can pass a tag or a list of tags.

In case the tag is duplicated nothing will happen.

It returns a list of associated tags

remove(struct, tag, context \\ "tags")

remove(struct, tag, context) :: tag_list

Get a persisted struct and removes the tag association for a specific context.

In case the association doesn’t exist nothing will happen.

It returns a list of associated tags

tag_list(struct, context \\ "tags")

tag_list(struct, context) :: tag_list

Given a struct, it searchs the associated tags for a specific context.

It returns a list of associated tags ordered by insert_date

tagged_with(tag, model, context \\ "tags")

tagged_with(tag, module, context) :: list

Given a tag, model and context (‘tag’ by default), will find all the model resources associated to the given tag.

tags(model, context \\ "tags")

tags(module, context) :: list

Search for all tags associated to a taggable_type and a context